+++ src/interval.rs
@@ -137,6 +137,7 @@ impl PartialOrd for Decoration {
 }
 
 #[derive(Clone, Copy, Debug)]
+#[repr(C)]
 pub struct DecoratedInterval {
     pub(crate) x: Interval,
     pub(crate) d: Decoration,
@@ -310,39 +311,61 @@ macro_rules! dec_interval {
     };
 }
 
-#[doc(hidden)]
-#[repr(C)]
-pub union _interval_rep {
-    pub f: [f64; 2],
-    pub i: Interval,
-}
-
 #[macro_export]
 macro_rules! const_interval {
     ($a:expr, $b:expr) => {{
-        const_assert!($a <= $b && $a != f64::INFINITY && $b != f64::NEG_INFINITY);
-        unsafe { $crate::_interval_rep { f: [-$a, $b] }.i }
+        #[repr(C)]
+        union Rep {
+            f: [f64; 2],
+            i: $crate::Interval,
+        }
+
+        static_assertions::const_assert!(
+            $a <= $b && $a != f64::INFINITY && $b != f64::NEG_INFINITY
+        );
+
+        #[allow(unused_unsafe)]
+        unsafe {
+            Rep { f: [-$a, $b] }.i
+        }
     }};
 }
 
 #[macro_export]
 macro_rules! const_dec_interval {
     ($a:expr, $b:expr) => {{
-        DecoratedInterval {
-            x: const_interval!($a, $b),
-            d: if $a == f64::NEG_INFINITY || $b == f64::INFINITY {
-                Decoration::Dac
-            } else {
-                Decoration::Com
-            },
+        #[derive(Clone, Copy)]
+        #[repr(C)]
+        struct _DecoratedInterval {
+            x: $crate::Interval,
+            d: $crate::Decoration,
+        }
+
+        #[repr(C)]
+        union Rep {
+            _di: _DecoratedInterval,
+            di: $crate::DecoratedInterval,
+        }
+
+        #[allow(unused_unsafe)]
+        unsafe {
+            Rep {
+                _di: _DecoratedInterval {
+                    x: const_interval!($a, $b),
+                    d: if $a == f64::NEG_INFINITY || $b == f64::INFINITY {
+                        $crate::Decoration::Dac
+                    } else {
+                        $crate::Decoration::Com
+                    },
+                },
+            }
+            .di
         }
     }};
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
     #[test]
     fn macros() {
         assert_eq!(interval!(1.0, 1.0).unwrap(), const_interval!(1.0, 1.0));